Refactor and reorganize profile-related classes.
Moved profile-specific classes (ProfileName, ProfileStatus, ProfilePicture, FetchProfile, and PrivacySettings) into separate files to improve maintainability and organization. Adjusted relevant imports and resolved minor inconsistencies in validation logic.
