+++ src/Models/ProfilePicture.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace SamuelTerra22\LaravelEvolutionClient\Models;
+
+class ProfilePicture extends Profile
+{
+    /**
+     * Create a new ProfilePicture instance.
+     *
+     * @param string $picture
+     */
+    public function __construct(string $picture)
+    {
+        parent::__construct(['picture' => $picture]);
+    }
+}
