+++ src/Models/ProfileStatus.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace SamuelTerra22\LaravelEvolutionClient\Models;
+
+class ProfileStatus extends Profile
+{
+    /**
+     * Create a new ProfileStatus instance.
+     *
+     * @param string $status
+     */
+    public function __construct(string $status)
+    {
+        parent::__construct(['status' => $status]);
+    }
+}
