+++ tests/Unit/Resources/InstanceResourceTest.php
@@ -25,7 +25,7 @@ class InstanceResourceTest extends TestCase
         $result = $this->instanceResource->getQrCode();
 
         $this->assertIsArray($result);
-        $this->assertEquals('success', $result['status']);
+        $this->assertEquals('connected', $result['status']);
         $this->assertArrayHasKey('qrcode', $result);
     }
